Change game instructions and win logic
he welcome message has been updated to indicate the spacebar as the input to play.
The starting amount has been increased from 20 to 100, and the bet is now 3.
The balance variable is now initialized with the value of CREDIT.
A new loop for checking horizontal wins has been added, which increases the balance on wins and decreases it on losses.
The loop condition has been adjusted to use the spacebar to continue the game.